+++ src/lib/gadgets/bitwise.unit-test.ts
@@ -57,27 +57,38 @@ let Bitwise = ZkProgram({
         return Gadgets.and(a, b, 64);
       },
     },
-    rot: {
+    rot32: {
       privateInputs: [Field],
       method(a: Field) {
-        return Gadgets.rotate(a, 12, 'left');
+        return Gadgets.rotate32(a, 12, 'left');
       },
     },
-    leftShift: {
+    rot64: {
       privateInputs: [Field],
       method(a: Field) {
-        return Gadgets.leftShift(a, 12);
+        return Gadgets.rotate64(a, 12, 'left');
       },
     },
-    rightShift: {
+    leftShift64: {
       privateInputs: [Field],
       method(a: Field) {
-        return Gadgets.rightShift(a, 12);
+        return Gadgets.leftShift64(a, 12);
+      },
+    },
+    leftShift32: {
+      privateInputs: [Field],
+      method(a: Field) {
+        return Gadgets.leftShift32(a, 12);
+      },
+    },
+    rightShift64: {
+      privateInputs: [Field],
+      method(a: Field) {
+        return Gadgets.rightShift64(a, 12);
       },
     },
   },
 });
-
 await Bitwise.compile();
 
 [2, 4, 8, 16, 32, 64, 128].forEach((length) => {
@@ -104,15 +115,26 @@ await Bitwise.compile();
 [2, 4, 8, 16, 32, 64].forEach((length) => {
   equivalent({ from: [uint(length)], to: field })(
     (x) => Fp.rot(x, 12n, 'left'),
-    (x) => Gadgets.rotate(x, 12, 'left')
+    (x) => Gadgets.rotate64(x, 12, 'left')
   );
   equivalent({ from: [uint(length)], to: field })(
     (x) => Fp.leftShift(x, 12),
-    (x) => Gadgets.leftShift(x, 12)
+    (x) => Gadgets.leftShift64(x, 12)
   );
   equivalent({ from: [uint(length)], to: field })(
     (x) => Fp.rightShift(x, 12),
-    (x) => Gadgets.rightShift(x, 12)
+    (x) => Gadgets.rightShift64(x, 12)
+  );
+});
+
+[2, 4, 8, 16, 32].forEach((length) => {
+  equivalent({ from: [uint(length)], to: field })(
+    (x) => Fp.rot(x, 12n, 'left', 32n),
+    (x) => Gadgets.rotate32(x, 12, 'left')
+  );
+  equivalent({ from: [uint(length)], to: field })(
+    (x) => Fp.leftShift(x, 12, 32),
+    (x) => Gadgets.leftShift32(x, 12)
   );
 });
 
@@ -167,7 +189,17 @@ await equivalentAsync({ from: [field], to: field }, { runs: 3 })(
     return Fp.rot(x, 12n, 'left');
   },
   async (x) => {
-    let proof = await Bitwise.rot(x);
+    let proof = await Bitwise.rot64(x);
+    return proof.publicOutput;
+  }
+);
+
+await equivalentAsync({ from: [uint(32)], to: uint(32) }, { runs: 30 })(
+  (x) => {
+    return Fp.rot(x, 12n, 'left', 32n);
+  },
+  async (x) => {
+    let proof = await Bitwise.rot32(x);
     return proof.publicOutput;
   }
 );
@@ -178,7 +210,18 @@ await equivalentAsync({ from: [field], to: field }, { runs: 3 })(
     return Fp.leftShift(x, 12);
   },
   async (x) => {
-    let proof = await Bitwise.leftShift(x);
+    let proof = await Bitwise.leftShift64(x);
+    return proof.publicOutput;
+  }
+);
+
+await equivalentAsync({ from: [field], to: field }, { runs: 3 })(
+  (x) => {
+    if (x >= 2n ** 64n) throw Error('Does not fit into 64 bits');
+    return Fp.leftShift(x, 12, 32);
+  },
+  async (x) => {
+    let proof = await Bitwise.leftShift32(x);
     return proof.publicOutput;
   }
 );
@@ -189,7 +232,7 @@ await equivalentAsync({ from: [field], to: field }, { runs: 3 })(
     return Fp.rightShift(x, 12);
   },
   async (x) => {
-    let proof = await Bitwise.rightShift(x);
+    let proof = await Bitwise.rightShift64(x);
     return proof.publicOutput;
   }
 );
@@ -229,6 +272,6 @@ let isJustRotate = ifNotAllConstant(
   and(contains(rotChain), withoutGenerics(equals(rotChain)))
 );
 
-constraintSystem.fromZkProgram(Bitwise, 'rot', isJustRotate);
-constraintSystem.fromZkProgram(Bitwise, 'leftShift', isJustRotate);
-constraintSystem.fromZkProgram(Bitwise, 'rightShift', isJustRotate);
+constraintSystem.fromZkProgram(Bitwise, 'rot64', isJustRotate);
+constraintSystem.fromZkProgram(Bitwise, 'leftShift64', isJustRotate);
+constraintSystem.fromZkProgram(Bitwise, 'rightShift64', isJustRotate);